pyproject.toml
[project]
name = "stereo-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]
